refactor: modify test Package Structure
