+++ src/test/java/racingcar/util/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
